+++ frontend/src/app/[locale]/subscribe/page.tsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 import pick from "lodash/pick";
 import { Metadata } from "next";
 import SubscriptionForm from "src/app/[locale]/subscribe/SubscriptionForm";
